backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.1.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
dnspython==1.16.0
ecdsa==0.19.1
fastapi==0.129.0
h11==0.16.0
idna==3.11
nba_api==1.11.4
numpy==2.4.2
pandas==3.0.1
passlib==1.7.4
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==3.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
